requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
gunicorn==21.2.0
mysqlclient==2.2.6
packaging==25.0
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
